+++ R/ggpar.R
@@ -9,7 +9,8 @@ NULL
 #'   palettes e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue",
 #'   "red"); and scientific journal palettes from ggsci R package, e.g.: "npg",
 #'   "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and
-#'   "rickandmorty".
+#'   "rickandmorty". Can be also a numeric vector of length(groups); in this
+#'   case a basic color palette is created using the function \link[grDevices](palette).
 #' @param gradient.cols vector of colors to use for n-colour gradient. Allowed
 #'   values include brewer and ggsci color palettes.
 #' @param main,title plot main title.
@@ -162,6 +163,7 @@ ggpar <- function(p, palette = NULL, gradient.cols = NULL,
   if(!is.null(subtitle)) submain <- subtitle
   if(!is.null(font.title)) font.main <- font.title
   if(!is.null(font.subtitle)) font.submain <- font.subtitle
+  if(is.numeric(palette)) palette <- grDevices::palette()[palette]
 
 
   for(i in 1:length(list.plots)){