+++ R/ggstripchart.R
@@ -20,7 +20,8 @@ NULL
 #'   palette = "Dark2" or palette = c("gray", "blue", "red") \item legend title,
 #'   labels and position: legend = "right" \item plot orientation : orientation
 #'   = c("vertical", "horizontal", "reverse") }
-#' @seealso \code{\link{ggpar}}
+#'@seealso \code{\link{ggpar}}, \code{\link{ggviolin}}, \code{\link{ggdotplot}}
+#'  and \code{\link{ggboxplot}}.
 #' @examples
 #' # Load data
 #' data("ToothGrowth")