+++ NEWS.md
@@ -5,6 +5,8 @@
 ## Major changes
    
 ## Minor changes
+
+- Now, the argument `palette` Can be also a numeric vector of length(groups); in this case a basic color palette is created using the function `grDevices::palette().
    
 # Bug fixes